+++ changelog.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * New option to Stop/Start only the triggers that are already Started #291
 ### Fixed
 * `RootFolder` must be absolute otherwise temp files cannot be written #335
+* Catch InvalidOperation exception when reading empty FilterFilePath in New-AdfPublishOption.ps1 #338
 
 ## [1.5.0] - 2023-06-18
 ### Added